Created OpeningSearch subcomponent
